+++ src/main/java/Application.java
@@ -1,3 +1,5 @@
+import lotto.controller.LottoController;
+
 public class Application {
     public static void main(String[] args) {
         LottoController lottoController = new LottoController();
